remove unused line/bar reading functions & compressed from file names
- Old SQL/functions related to old way to doing readings was removed.
A new migration for dropping the removed functions was added. Another
migration to drop old views was added.
- remove compressed from all file names except test (coming)
and migrations. Moved what was in compressed-readings.ts
into existing readings.ts.
- Noted test files that probably need to rename when tests updated.
- small fixes
