+++ src/client/app/containers/ExportContainer.ts
@@ -6,10 +6,8 @@ import { connect } from 'react-redux';
 import * as _ from 'lodash';
 import ExportComponent from '../components/ExportComponent';
 import { ChartTypes } from '../types/redux/graph';
-import { ExportDataSet } from '../types/readings';
+import { CompressedBarReading, CompressedLineReading, ExportDataSet } from '../types/readings';
 import { State } from '../types/redux/state';
-import { CompressedBarReading, CompressedLineReading } from '../types/compressed-readings';
-
 
 function transformLineReadingToLegacy(reading: CompressedLineReading): [number, number, number] {
 	return [reading.startTimestamp, reading.reading, reading.endTimestamp];
