+++ src/server/models/Reading.js
@@ -33,48 +33,13 @@ class Reading {
 	}
 
 	/**
-	 * Returns a promise to create the compressed readings function.
-	 * @param conn the database connection to use
-	 * @return {Promise.<>}
-	 */
-	static createCompressedReadingsFunction(conn) {
-		return conn.none(sqlFile('reading/create_function_get_compressed_readings.sql'));
-	}
-
-	/**
-	 * Returns a promise to create the compressed groups readings function.
-	 * @param conn the database connection to use
-	 * @return {Promise.<>}
-	 */
-	static createCompressedGroupsReadingsFunction(conn) {
-		return conn.none(sqlFile('reading/create_function_get_compressed_groups_readings.sql'));
-	}
-
-	/**
-	 * Returns a promise to create the compressed groups barchart readings function.
-	 * @param conn the database connection to use
-	 * @return {Promise.<>}
-	 */
-	static createCompressedGroupsBarchartReadingsFunction(conn) {
-		return conn.none(sqlFile('reading/create_function_get_group_barchart_readings.sql'));
-	}
-
-	/**
-	 * Returns a promise to create the barchart readings function.
-	 * @return {Promise.<>}
-	 */
-	static createBarchartReadingsFunction(conn) {
-		return conn.none(sqlFile('reading/create_function_get_barchart_readings.sql'));
-	}
-
-	/**
 	 * Returns a promise to create the function and materialized views that aggregate
 	 * readings by various time intervals.
 	 * @param conn the database connection to use
 	 * @return {Promise<void>}
 	 */
 	static createCompressedReadingsMaterializedViews(conn) {
-		return conn.none(sqlFile('reading/create_compressed_reading_views.sql'));
+		return conn.none(sqlFile('reading/create_reading_views.sql'));
 	}
 
 	/**
@@ -239,108 +204,6 @@ class Reading {
 	 */
 	insertOrIgnore(conn) {
 		return conn.none(sqlFile('reading/insert_or_ignore_reading.sql'), this);
-	}
-
-	/**
-	 * Gets a number of compressed readings that approximate the given time range for the given meters.
-	 *
-	 * Compressed readings are in kilowatts.
-	 * @param meterIDs an array of ids for meters whose points are being compressed
-	 * @param {Moment?} fromTimestamp An optional start point for the time range.
-	 * @param {Moment?} toTimestamp An optional end point for the time range
-	 * @param numPoints The number of points to compress to. Defaults to 500
-	 * @param conn is the connection to use.
-	 * @return {Promise<object<int, array<{reading_rate: number, start_timestamp: Date, end_timestamp: Date}>>>}
-	 */
-	static async getCompressedReadings(meterIDs, fromTimestamp = null, toTimestamp = null, numPoints = 500, conn) {
-		const allCompressedReadings = await conn.func('compressed_readings',
-			[meterIDs, fromTimestamp || '-infinity', toTimestamp || 'infinity', numPoints]);
-		// Separate the result rows by meter_id and return a nested object.
-		const compressedReadingsByMeterID = mapToObject(meterIDs, () => []); // Returns { 1: [], 2: [], ... }
-		// For each row in the allCompressedReadings table, append the compressed reading value to the array for
-		// the meter that corresponds to that reading.
-		for (const row of allCompressedReadings) {
-			compressedReadingsByMeterID[row.meter_id].push(
-				{ reading_rate: row.reading_rate, start_timestamp: row.start_timestamp, end_timestamp: row.end_timestamp }
-			);
-		}
-		return compressedReadingsByMeterID;
-	}
-
-	/**
-	 * Gets a number of compressed readings that approximate the given time range for the given groups.
-	 *
-	 * Compressed readings are in kilowatts.
-	 * @param groupIDs an array of ids for groups whose points are being compressed
-	 * @param {Moment?} fromTimestamp An optional start point for the time range.
-	 * @param {Moment?} toTimestamp An optional end point for the time range
-	 * @param numPoints The number of points to compress to. Defaults to 500
-	 * @param conn is the connection to use.
-	 * @return {Promise<object<int, array<{reading_rate: number, start_timestamp: Moment, end_timestamp: Moment}>>>}
-	 */
-	static async getCompressedGroupReadings(groupIDs, fromTimestamp = null, toTimestamp = null, numPoints = 500, conn) {
-		const allCompressedReadings = await conn.func('compressed_group_readings',
-			[groupIDs, fromTimestamp || '-infinity', toTimestamp || 'infinity', numPoints]);
-		// Separate the result rows by meter_id and return a nested object.
-		const compressedReadingsByGroupID = mapToObject(groupIDs, () => []); // Returns { 1: [], 2: [], ... }
-		for (const row of allCompressedReadings) {
-			compressedReadingsByGroupID[row.group_id].push(
-				{ reading_rate: row.reading_rate, start_timestamp: row.start_timestamp, end_timestamp: row.end_timestamp }
-			);
-		}
-		return compressedReadingsByGroupID;
-	}
-
-	/**
-	 * Gets barchart readings for every meter across the given time interval for a duration.
-	 *
-	 * Compressed readings are in kilowatts.
-	 * @param meterIDs an array of ids for meters whose points are being compressed
-	 * @param duration A moment time duration over which to sum the readings
-	 * @param fromTimestamp An optional start point for the beginning of the entire time range.
-	 * @param toTimestamp An optional end point for the end of the entire time range.
-	 * @param conn is the connection to use.
-	 * @return {Promise<object<int, array<{reading_sum: number, start_timestamp: Date, end_timestamp: Date}>>>}
-	 */
-	static async getBarchartReadings(meterIDs, duration, fromTimestamp = null, toTimestamp = null, conn) {
-		const allBarchartReadings = await conn.func('barchart_readings', [meterIDs, duration, fromTimestamp || '-infinity', toTimestamp || 'infinity']);
-		// Separate the result rows by meter_id and return a nested object.
-		const barchartReadingsByMeterID = mapToObject(meterIDs, () => []);
-		for (const row of allBarchartReadings) {
-			barchartReadingsByMeterID[row.meter_id].push(
-				{ reading_sum: row.reading_sum, start_timestamp: row.start_timestamp, end_timestamp: row.end_timestamp }
-			);
-		}
-		return barchartReadingsByMeterID;
-	}
-
-	/**
-	 * Gets barchart readings for every group across the given time interval for a duration.
-	 *
-	 * Compressed readings are in kilowatts.
-	 * @param groupIDs an array of ids for groups whose points are being compressed
-	 * @param duration A moment time duration over which to sum the readings
-	 * @param fromTimestamp An optional start point for the beginning of the entire time range.
-	 * @param toTimestamp An optional end point for the end of the entire time range.
-	 * @param conn is the connection to use.
-	 * @return {Promise<object<int, array<{reading_sum: number, start_timestamp: Date, end_timestamp: Date}>>>}
-	 */
-	static async getGroupBarchartReadings(groupIDs, duration, fromTimestamp = null, toTimestamp = null, conn) {
-		const allBarchartReadings = await conn.func('barchart_group_readings',
-			[groupIDs, duration, fromTimestamp || '-infinity', toTimestamp || 'infinity']);
-		// Separate the result rows by meter_id and return a nested object.
-		const barchartReadingsByGroupID = mapToObject(groupIDs, () => []);
-		for (const row of allBarchartReadings) {
-			// If there was an unexpected group id in the barchart readings by group, just create a place to hold
-			// that value in the object to be returned.
-			if (barchartReadingsByGroupID[row.group_id] === undefined) {
-				barchartReadingsByGroupID[row.group_id] = [];
-			}
-			barchartReadingsByGroupID[row.group_id].push(
-				{ reading_sum: row.reading_sum, start_timestamp: row.start_timestamp, end_timestamp: row.end_timestamp }
-			);
-		}
-		return barchartReadingsByGroupID;
 	}
 
 	/**
