+++ src/server/models/database.js
@@ -95,17 +95,12 @@ async function createSchema(conn) {
 	await Meter.createMeterTypesEnum(conn);
 	await Meter.createTable(conn);
 	await Reading.createTable(conn);
-	await Reading.createCompressedReadingsFunction(conn);
-	await Reading.createCompressedGroupsReadingsFunction(conn);
-	await Reading.createBarchartReadingsFunction(conn);
-	await Reading.createCompressedGroupsBarchartReadingsFunction(conn);
 	await User.createUserTypesEnum(conn);
 	await User.createTable(conn);
 	await Preferences.createTable(conn);
 	await Group.createTables(conn);
 	await Migration.createTable(conn);
 	await LogEmail.createTable(conn);
-	await conn.none(sqlFile('reading/create_function_get_compressed_readings.sql'));
 	await Reading.createCompressedReadingsMaterializedViews(conn);
 	await Reading.createCompareReadingsFunction(conn);
 	await Baseline.createTable(conn);
